% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CytOpt.R
\name{plot.CytOpt}
\alias{plot.CytOpt}
\title{CytOpt plot}
\usage{
\method{plot}{CytOpt}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CytOpt} to plot.}

\item{...}{further arguments passed to or from other methods. Not implemented.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object, potentially composed through 
\code{\link[patchwork]{patchwork}}
}
\description{
plot S3 method for CytOpt object
}
\examples{
if(interactive()){

res <- CytOpT(X_s = HIPC_Stanford_1228_1A, X_t = HIPC_Stanford_1369_1A, 
             Lab_source = HIPC_Stanford_1228_1A_labels,
             eps = 0.0001, lbd = 0.0001, n_iter = 10000, n_stoc=10,
             step_grad = 10, step = 5, power = 0.99, 
             method='minmax')
plot(res)

}
}
