% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CytOpt.R
\name{print.CytOpt}
\alias{print.CytOpt}
\title{CytOpt print}
\usage{
\method{print}{CytOpt}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CytOpt} to print.}

\item{...}{further arguments passed to or from other methods. Not implemented.}
}
\value{
the proportions \code{data.frame} from \code{x}
}
\description{
print S3 method for CytOpt object
}
\examples{
if(interactive()){

res <- CytOpT(X_s = HIPC_Stanford_1228_1A, X_t = HIPC_Stanford_1369_1A, 
             Lab_source = HIPC_Stanford_1228_1A_labels,
             eps = 0.0001, lbd = 0.0001, n_iter = 10000, n_stoc=10,
             step_grad = 10, step = 5, power = 0.99, 
             method='minmax')
print(res)

}
}
