% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CytOpt.R
\name{print.summary.CytOpt}
\alias{print.summary.CytOpt}
\title{CytOpt print summary}
\usage{
\method{print}{summary.CytOpt}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.CytOpt} to print.}

\item{...}{further arguments passed to or from other methods. Not implemented.}
}
\description{
print S3 method for summary.CytOpt object
}
